Update readmes, rename files in backend
