+++ backend/routers/recipe.py
@@ -6,9 +6,9 @@ from fastapi import Request
 from fastapi import APIRouter
 
 from recipedex import App
-from backend.database.model import ResponseModel
-from backend.database.database import retrieve_recipes
-from backend.database.database import add_recipe
+from backend.data.model import ResponseModel
+from backend.data.database import add_recipe
+from backend.data.database import get_recipes
 
 
 logger = logging.getLogger("backend.api.routers.recipes")
@@ -21,8 +21,8 @@ router = APIRouter(
 
 
 @router.get("/all", response_description="Retrieve all scraped urls")
-async def get_recipes():
-    recipes = await retrieve_recipes()
+async def all():
+    recipes = await get_recipes()
     if recipes:
         return ResponseModel(recipes, "Recipe data retrieved successfully")
     return ResponseModel(recipes, "Empty list returned")
